Tests/LibWeb/Text/input/css/style-invalidation/sheet-add-first-child-pseudo-only.html

<!DOCTYPE html>
<script src="../../include.js"></script>
<div id="container">
    <div class="kid">first</div>
    <div class="kid">second</div>
    <div class="kid">third</div>
</div>
<script>
    function settleAndReset(triggerElement) {
        getComputedStyle(triggerElement).color;
        getComputedStyle(triggerElement).color;
        internals.resetStyleInvalidationCounters();
    }

    function addBystanders(parent, count) {
        for (let i = 0; i < count; ++i) {
            const bystander = document.createElement("div");
            bystander.textContent = `bystander ${i}`;
            parent.appendChild(bystander);
        }
    }

    function dumpCounters(label) {
        const c = internals.getStyleInvalidationCounters();
        println(`${label}: fullStyleInvalidations=${c.fullStyleInvalidations}, styleInvalidations=${c.styleInvalidations}, elementStyleRecomputations=${c.elementStyleRecomputations}, elementStyleNoopRecomputations=${c.elementStyleNoopRecomputations}`);
    }

    test(() => {
        const container = document.getElementById("container");
        addBystanders(document.body, 25);
        const firstKid = container.firstElementChild;
        settleAndReset(firstKid);

        // Adding a sheet whose only rule's rightmost compound is :first-child
        // (no class/tag/id/attr feature) should target boundary children only,
        // not fall back to a whole-subtree invalidation that recomputes every
        // element in the document.
        const firstChildStyle = document.createElement("style");
        firstChildStyle.textContent = ":first-child { color: rgb(0, 0, 128); }";
        document.head.appendChild(firstChildStyle);
        getComputedStyle(firstKid).color;

        dumpCounters(":first-child-only sheet add");
    });
</script>
