Tests/LibWeb/Text/expected/css/style-invalidation/sheet-add-first-child-pseudo-only.txt

:first-child-only sheet add: fullStyleInvalidations=1, styleInvalidations=3, elementStyleRecomputations=36, elementStyleNoopRecomputations=0
